Refactor and add interfaces for managing circular deps
